+++ web-common/src/components/menu/DashboardDraggableList.svelte
